refactor(project): move useFrozenWhileHidden into its own hook file
