+++ app/src/hooks/index.ts
@@ -4,6 +4,7 @@ export { useLocale } from "react-aria-components";
 // Custom hooks
 export * from "./useDebouncedChange";
 export * from "./useDeferredVisibility";
+export * from "./useFrozenWhileHidden";
 export * from "./useDebouncedJSONSync";
 export * from "./useInterval";
 export * from "./useWordColor";
